functions to responseUtils
